feat(driver): standalone driver init + structure
